Two emoji and the word checkpoint
